split Filters and Sorts to their own components
